examples/requirements.txt
pyjwt==1.7.1
paho-mqtt==1.4.0
cryptography==2.7
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-pubsub==0.45.0
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
google-cloud==0.34.0
